chore: added CODEOWNERS file (#11)
This file contains the usernames of the code's owners. The file syntax
is similar to that of .gitignore. The owners will be automatically added
to the review line of PRs. This should make it easier for us to request
PR reviews, and it will definitely be easier for OSS contributors who
want to send a PR.

More info: https://help.github.com/en/github/creating-cloning-and-archiving-repositories/about-code-owners