feat: implement development and runtime images (#43)
Add a `Dockerfile` and some support to generate development and runtime
images to build and deploy applications based on the framework. The
tests are very minimal at this point, we just check that the images can
be created, we do not verify that they are useful in any way. Additional
tests will come in future PRs.

The `Dockerfile` has a number of weird looking stages to help me debug it
and incrementally build the cache. I think they are no longer needed,
but they do not hurt, and may be needed again in the near future (if
something breaks).

We also will need some scripts to use these images, and some integration
tests, but we need to start somewhere.