ci: use previous cache when vcpkg.json changes
TIL: a bit more about how GitHub Actions caching works. On a perfect
match for the hashing key the cache is not updated. When the inputs to
the cache change you certainly want to update it, but maybe you want to
reuse the previous version to start with a partially warm cache.  That
is what `restore-keys` is for. Warms up the cache, even if the inputs
change, and then update with the result.