cleanup: simplify HttpRequest implementation
We originally thought that this class would need a `PImpl` to hide the
Boost.Beast implementation details. That idea failed when we realized
that Boost.Beast returns `std::string_view` and we did not want to do
so. Effectively this meant we were copying (or moving, when possible)
the data into our own data structure, and then using a `PImpl` without
any benefits.