cleanup: improve testing for non-site examples
We have some examples that are specific to the framework, and do not
follow the region tags or any other conventions from the cloud site.
These were not tested, now they are.

The GCS example needs an integration test, but we have not figured out
how to do that just yet.