ci: include vcpkg in vcpkg cache hash (#182)
It seems that the builders for GitHub actions get updated from time to
time. With a new vcpkg binary all the existing vcpkg binary caches
become invalid, but GitHub actions did not know that, and keeps trying
to use it.

Also, we can share the vcpkg cache across builds, so let's do that.