feat: fix compilation problems on MSVC (#233)
Most of the problems were around `setenv()` and `getenv()`, which I
solved by writing a small wrappers. Kept `std::getenv()` in the
examples, because that is most familiar to C++ developers, had to turn
off warnings from MSVC about it though.

Avoid googletest regex because they are unsupported with MSVC

Fix executable paths in integration tests