doc: improve local development how-to guide
The command to bootstrap `vcpkg` was wrong. But more importantly, we did
not explain in detail what was the expected output and results from each
command. In some cases I added options to each command to reduce the
informational output.

I also simplified the test function, there is no need to have a clean
shutdown, and having that code there is distracting.