chore: update to pack v0.17.0
Use pack v0.17.0 in the CI builds and the examples, this is the current
version and has slightly different syntax for some commands than what we
were showing customers in the how-to guides. Also, I think it makes
sense to run CI with the same version we use for the how-to guides.