ci: cleanup images created during builds
This change reduces the number of images pushed to Google Container
Registry, and removes the images created during the build.

If the build crashes, the cleanup step does not get to run, so as early
as possible a background step (one that has no dependencies in the
build and always succeeds) finds any images older than 4 weeks and
deletes them.