cleanup: split CI vs. developer builders (#265)
In the CI builds we need to test how the *current* version of
`functions-framework-cpp` works with buildback builders. But users of
the framework can use the *released* version of
`functions-framework-cpp`, as this is the version they get for local
development too. This PR splits the two builders, using different images
and buildpacks for each.