refactor: use generator to create buildpacks
Buildpack builders are not very configurable (by design), and hardcode
the name of the build and run images in their definition files. We also
have 3 different builders that we use in different cases:

- One is the builder we recommend for users, when they create the run
  and build images locally. This builder uses the latest version of the
  framework.
- Another is the builder we recommend for users, when they create the
  run and build images with Cloud Build. This needs diferent names for
  these images, depending on the project used by the customer.
- Finally the builder used in CI builds, that uses the *current* version
  of the framework.

This PR introduces a shell script to create these builders. It also
changes the CI scripts to verify the builders are kept up to date. To
ease the code generation I changed the image names to always be of the
form: `${prefix}run-image` and `${prefix}build-image`. That required
additional changes to the how-to guides and miscellaneous README files.