feat: make builder self-contained
With this change the builder does not depend on having particular files
installed in the builder image. That means the builder image only needs
to have packages installed via `apt`, and we can evolve the builder
separately from the builder image (at least when not including binary
caches in the builder image).