cleanup: remove unused buildpack (#344)
We used to create a buildpack for the CI builds, and a separate
buildpack for the user-facing READMEs. A while back we changed the
user-facing READMEs to use the official GCP buildpacks, this cleans up
the now (largely) unused code.

I also updated the documents that referenced these buildpacks when they
shouldn't, and I took the opportunity to (a) clarify what is the purpose
and intended audience of these docs, (b) cross-reference the docs in
case folks land in the wrong document, and (c) create an index of the
key how-to guides to have a single thing to cross-reference.